# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: M2 - Atención Ciudadana API
  description: API REST del módulo M2 de Atención Ciudadana, Reclamos y Solicitudes.
  version: 1.0.0
servers:
  - url: /
    description: Mismo origen que sirve Swagger UI
...
paths:
...
  /api/tickets:
    get:
      tags: [ Tickets ]
      summary: Listar tickets con filtros
      description: |
        Bandeja de tickets para agentes/administradores, con filtros combinables
        por categoría, prioridad, barrio, área responsable y estado (Story 3.1).
      operationId: listTickets
      security:
        - bearerAuth: [ ]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: priority
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Priority'
        - name: neighborhoodId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: responsibleAreaId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: sort
          in: query
          required: false
          description: Propiedades ordenables de Ticket (whitelist propia del backend); formato `campo,direccion`.
          schema:
            type: array
            items:
              type: string
          example: [ "createdAt,desc" ]
      responses:
        '200':
          description: Página de tickets que cumplen los filtros.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedTicketResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Tickets]
      summary: Crear un ticket
      description: |
        Crea un ticket autenticado. La misma operación acepta JSON puro o multipart.

        En JSON no se adjuntan archivos; si el riesgo calculado internamente es HIGH o CRITICAL,
        responde 422 EVIDENCE_REQUIRED. En multipart, `data` es un part JSON obligatorio y
        `evidence` admite hasta 5 archivos opcionales, de 10 MB cada uno y 50 MB en total.
        MIME admitidos: image/jpeg, image/png, image/webp, image/heic, image/heif y application/pdf.
      operationId: createTicket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
          multipart/form-data:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  $ref: '#/components/schemas/CreateTicketRequest'
                evidence:
                  type: array
                  maxItems: 5
                  description: Archivos opcionales; máximo 10 MB por archivo y 50 MB en total.
                  items:
                    type: string
                    format: binary
            encoding:
              data:
                contentType: application/json
              evidence:
                contentType: image/jpeg, image/png, image/webp, image/heic, image/heif, application/pdf
      responses:
        '201':
          description: Ticket creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTicketResponse'
        '400':
          description: Request inválido, part `data` ausente o adjunto inválido. INVALID_ATTACHMENT sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                missingData:
                  value:
                    code: INVALID_REQUEST
                    message: Falta el part obligatorio 'data'
                invalidAttachment:
                  value:
                    code: INVALID_ATTACHMENT
                    message: Cada evidencia debe contener un archivo no vacío.
        '401':
          $ref: '#/components/responses/InvalidToken'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: Límite de adjuntos excedido; sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: PAYLOAD_TOO_LARGE
                message: La evidencia supera el tamaño máximo permitido.
        '415':
          description: Content-Type del request no soportado o MIME de evidencia no admitido. El MIME de evidencia sólo aplica a multipart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: UNSUPPORTED_MEDIA_TYPE
                message: Tipo de contenido no soportado
        '422':
          description: El riesgo calculado requiere al menos una evidencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: EVIDENCE_REQUIRED
                message: El nivel de riesgo detectado requiere adjuntar al menos una evidencia.
        '503':
          description: El almacenamiento de adjuntos no está disponible; sólo aplica cuando multipart contiene evidencia.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: ATTACHMENT_STORAGE_UNAVAILABLE
                message: No fue posible almacenar la evidencia. Intentá nuevamente más tarde.
  /api/tickets/{ticketId}:
    get:
      tags: [ Tickets ]
      summary: Consultar el detalle de un ticket propio
      description: |
        GET /tickets/{id} (Entidades V1.49 §"Autenticación, roles y vistas"):
        detalle ciudadano de un ticket, sólo para su propietario autenticado.
        Un ticket anónimo no tiene citizenId, así que nunca es "propio" de
        nadie acá; su consulta pasa por POST /tracking/access.
      operationId: getTicketById
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Detalle del ticket propio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/staff/tickets/{ticketId}:
    get:
      tags: [ Tickets ]
      summary: Consultar el detalle staff de un ticket
      description: |
        GET /staff/tickets/{id} (Guía funcional M2 §7/§7.1): detalle de
        gestión interna. AGENT y ADMIN acceden a cualquier ticket ajeno sin
        restricción de área; AREA_RESPONSIBLE sólo a los de su propia areaId
        (o al suyo propio como ciudadano). El endpoint citizen-view expone
        en cambio la misma proyección pública que recibe el propietario.
      operationId: getStaffTicketDetail
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Detalle staff del ticket.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetailResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/tickets/{ticketId}/review:
    post:
      tags: [ Tickets ]
      summary: Tomar un ticket a revisión
      description: |
        Transición REGISTERED -> IN_REVIEW; asigna el agente autenticado como
        assignedAgent si el ticket todavía no tiene uno asignado. Requiere rol
        AGENT o ADMIN y no permite usar capacidad administrativa sobre un ticket
        propio (Story 3.2).
      operationId: startTicketReview
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Ticket en revisión.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketStateConflict'
  /api/tickets/{ticketId}/classification:
    patch:
      tags: [ Tickets ]
      summary: Corregir la clasificación de un ticket
      description: |
        Permite reclasificar el RequestType durante la primera revisión del
        ticket (antes de derivarlo a un área), recalculando área responsable,
        cantidad estimada de afectados, plantilla de formulario, prioridad y
        el SLA inicial (Story 3.2). Requiere rol AGENT o ADMIN y no permite
        usar capacidad administrativa sobre un ticket propio.
      operationId: correctTicketClassification
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificationCorrectionRequest'
      responses:
        '200':
          description: Clasificación corregida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketStateConflict'
  /api/tickets/{ticketId}/route:
    post:
      tags: [ Tickets ]
      summary: Finalizar clasificación y comenzar o derivar la gestión
      description: |
        Requiere rol AGENT o ADMIN y no permite usar capacidad administrativa
        sobre un ticket propio. Si responsibleAreaId=M2, realiza la transición
        IN_REVIEW -> IN_PROGRESS e inicia la gestión propia sin evento ROUTED.
        Si el área es externa, realiza IN_REVIEW -> ROUTED y publica el snapshot
        ticketUpdated/ROUTED (Story 3.3 / DDA2-59).
      operationId: routeTicketToArea
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Clasificación finalizada y gestión iniciada o derivada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketStateConflict'
  /api/tickets/{ticketId}/information-request:
    post:
      tags: [Information Requests]
      summary: Solicitar información adicional
      description: Operación permitida a AGENT o ADMIN sobre un ticket ajeno y en un estado compatible.
      operationId: requestTicketInformation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInformationRequest'
      responses:
        '201':
          description: Solicitud de información creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationRequestResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InformationRequestConflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /api/tickets/{ticketId}/information-response:
    post:
      tags: [Information Requests]
      summary: Responder una solicitud de información
      description: Operación permitida al ciudadano propietario del ticket identificado.
      operationId: answerTicketInformation
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerInformationRequest'
      responses:
        '200':
          description: Solicitud de información respondida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InformationRequestResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InformationRequestConflict'
        '410':
          description: El plazo para responder venció.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                code: INFORMATION_REQUEST_EXPIRED
                message: El plazo para responder la solicitud de información venció
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /api/tickets/{ticketId}/resolution:
    post:
      tags: [ Tickets ]
      summary: Registrar la resolución manual de un ticket
      description: Operación de AGENT o ADMIN para tickets ajenos gestionados por M2 en estado IN_PROGRESS.
      operationId: resolveTicketManually
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveTicketRequest'
      responses:
        '201':
          description: Resolución registrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResolutionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketResolutionConflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /api/tickets/{ticketId}/resolution/confirm:
    post:
      tags: [ Tickets ]
      summary: Confirmar la resolución de un ticket
      description: Operación permitida al propietario de un ticket en estado RESOLVED.
      operationId: confirmTicketResolution
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: Resolución confirmada y ticket cerrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResolutionActionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketResolutionConflict'
  /api/tickets/{ticketId}/resolution/reopen:
    post:
      tags: [ Tickets ]
      summary: Reabrir un ticket por disconformidad
      description: Operación permitida al propietario de un ticket en estado RESOLVED.
      operationId: reopenTicketResolution
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReopenTicketRequest'
      responses:
        '200':
          description: Ticket reabierto para retomar su tratamiento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResolutionActionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketResolutionConflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /api/tickets/{ticketId}/cancel:
    post:
      tags: [ Tickets ]
      summary: Cancelar un ticket antes de su derivación
      description: |
        El ciudadano propietario sólo puede cancelar su propio ticket mientras
        está en REGISTERED, independientemente de su role almacenado. AGENT y
        ADMIN pueden cancelar administrativamente tickets ajenos en REGISTERED,
        IN_REVIEW o PENDING_INFORMATION. AREA_RESPONSIBLE no posee capacidad
        administrativa de cancelación. Un ticket ya derivado
        (ROUTED/IN_PROGRESS) se cancela por el flujo de integración
        (updateTicketStatus/REJECTED), no por este endpoint.
      operationId: cancelTicket
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTicketRequest'
      responses:
        '200':
          description: Ticket cancelado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidToken'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/TicketStateConflict'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token Bearer obtenido del proveedor de identidad vigente. En DEV puede obtenerse mediante /api/auth/login.
  parameters:
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    InvalidRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_REQUEST
            message: El request es inválido
    InvalidToken:
      description: Token Bearer ausente o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: INVALID_TOKEN
            message: La sesión no es válida
    Forbidden:
      description: La identidad autenticada no tiene permiso para realizar la operación.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: FORBIDDEN
            message: No está autorizado para realizar esta operación
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: NOT_FOUND
            message: Recurso no encontrado
...
    TicketStateConflict:
      description: El estado actual del ticket no permite la transición solicitada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            code: TICKET_STATE_CONFLICT
            message: El ticket está en un estado que no permite esta operación
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
...
    CreateTicketRequest:
      type: object
      required: [requestTypeId, summary, description, formData]
      properties:
        requestTypeId:
          type: integer
          format: int64
        summary:
          type: string
          maxLength: 200
        description:
          type: string
        formData:
          type: object
          description: Objeto dinámico cuyas keys corresponden al `code` de los campos del formulario vigente.
          additionalProperties: true
        location:
          allOf:
            - $ref: '#/components/schemas/LocationData'
          nullable: true
      additionalProperties: false
...
    TicketDetailResponse:
      type: object
      description: |
        Detalle plano del ticket. Conserva los campos útiles de TicketResponse,
        agrega descripción, metadatos de adjuntos e historial y expone sólo el
        nombre del barrio. En detalle ciudadano y citizen-view los adjuntos son
        PUBLIC y las actividades usan la proyección pública sanitizada; el detalle
        staff puede incluir adjuntos INTERNAL y campos adicionales del historial.
      required: [id, publicId, requestTypeCode, requestTypeName, categoryName, subcategoryName,
                 ticketType, summary, description, currentStatus, currentPriority, responsibleAreaId,
                 assignedAgentId, anonymous, estimatedAffectedCount, escalated, escalationReasonCode,
                 escalatedAt, firstResponseDueAt, firstResponseNearDue, firstResponseBreached,
                 resolutionDueAt, slaNearDue, slaBreached, resolutionNearDueAt, neighborhoodName,
                 classificationFinalizedAt, statusChangedAt, createdAt, updatedAt, attachments,
                 ticketActivities, labels]
      properties:
        id:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        requestTypeCode:
          type: string
        requestTypeName:
          type: string
        categoryName:
          type: string
        subcategoryName:
          type: string
        ticketType:
          $ref: '#/components/schemas/TicketType'
        summary:
          type: string
        description:
          type: string
        currentStatus:
          $ref: '#/components/schemas/TicketStatus'
        currentPriority:
          $ref: '#/components/schemas/Priority'
        responsibleAreaId:
          type: string
        assignedAgentId:
          type: string
          nullable: true
        anonymous:
          type: boolean
        estimatedAffectedCount:
          type: integer
          format: int32
        escalated:
          type: boolean
        escalationReasonCode:
          allOf:
            - $ref: '#/components/schemas/EscalationReasonCode'
          nullable: true
        escalatedAt:
          type: string
          format: date-time
          nullable: true
        firstResponseDueAt:
          type: string
          format: date-time
          nullable: true
        firstResponseNearDue:
          type: boolean
        firstResponseBreached:
          type: boolean
        resolutionDueAt:
          type: string
          format: date-time
          nullable: true
        slaNearDue:
          type: boolean
        slaBreached:
          type: boolean
        resolutionNearDueAt:
          type: string
          format: date-time
          nullable: true
        neighborhoodName:
          type: string
          nullable: true
        classificationFinalizedAt:
          type: string
          format: date-time
          nullable: true
        statusChangedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        attachments:
          type: array
          description: Metadatos únicamente; este contrato no provee descarga del archivo.
          items:
            $ref: '#/components/schemas/TicketAttachmentResponse'
        ticketActivities:
          type: array
          description: Historial ordenado por sequence ascendente.
          items:
            $ref: '#/components/schemas/TicketActivityResponse'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelSummaryResponse'
...
    PagedTicketResponse:
      type: object
      description: |
        Forma simplificada de la página que devuelve Spring Data (`Page<TicketResponse>`)
        para el listado de tickets; se documentan los campos estables para el
        cliente, sin los metadatos internos de paginación de Spring (`pageable`, `sort`).
      required: [content, totalElements, totalPages, size, number, first, last, numberOfElements, empty]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TicketResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
          description: Página actual (0-indexed).
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    ClassificationCorrectionRequest:
      type: object
      required: [requestTypeId]
      properties:
        requestTypeId:
          type: integer
          format: int64
      additionalProperties: false
    CreateTicketResponse:
      type: object
      required: [ticketId, publicId, trackingCode, status]
      properties:
        ticketId:
          type: string
          format: uuid
        publicId:
          $ref: '#/components/schemas/TicketPublicId'
        trackingCode:
          type: string
          description: Código que permite consultar el ticket mediante el endpoint público de tracking.
        status:
          $ref: '#/components/schemas/TicketStatus'
...
    TicketStatus:
      type: string
      enum: [REGISTERED, IN_REVIEW, ROUTED, IN_PROGRESS, PENDING_INFORMATION, DUPLICATE, RESOLVED, CLOSED, CANCELLED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
...